Fix installing APK wihout proper libs
Example:
The apk "soundhound" has support for x86, but the apk ask for some arm
libs even if it is installed with x86 libs.

At the moment Android wants to copy the ARM lib it will search for the
same lib (name) in the package, but then for x86. If it finds a lib with
the same name and for x86, it won't copy over the ARM lib.

This still needs some refinement. At the moment it says that it installs
with primary abi, but it installs with primary and secondary abi. This
needs testing.

Change-Id: Id80315a2136a58b4724c5c69ad92b5537ad334f4