App installed with Houdini now gets ARM arch
Some apps don't use the ABI and ABI2, but insteed are using the os.arch
property to know on which ARCH they are on. As for instance the HERE app
(2014) ask if the os.arch is 'arm'. If it is not, the app doesn't work.

This fix fools oly the apps that are installed with houdini support, so for
arm architecture, that they are on 'ARM' if asked with
getProperty(os.arch).

Change-Id: Ib272230917f38c03727ce05d1326660c2fea1532